+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `[mathutil]` Sub-package migrated to generics
 * `[sliceutil]` Sub-package migrated to generics
+* `[color]` Code refactoring
 
 ### 12.71.0