+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 12.72.0
+
+* `[terminal]` Added support of options for panels
+* `[mathutil]` Sub-package migrated to generics
+* `[sliceutil]` Sub-package migrated to generics
+* `[color]` Code refactoring
+* `[spinner]` Code refactoring
+
 ### 12.71.0
 
 * `[terminal]` Added panel size configuration feature